feat: support self hosted urls for migration commands
